+++ pytsite/fb/ep.py
@@ -9,9 +9,14 @@ __license__ = 'MIT'
 
 
 def authorize(args: dict, inp: dict):
-    session = _AuthSession(inp.get('state'))
+    """Authorization endpoint.
+    """
+    # Checking for errors
     error = inp.get('error_description')
     if error:
         _router.session.add_error(error)
+
+    # initializing authorization session
+    session = _AuthSession(inp.get('state'))
 
     return _http.response.Redirect(_router.url(session.final_redirect_uri, query=inp))